+++ src/grocery-button.js
@@ -1,37 +1,44 @@
-import React from 'react'
-import { withStyles, MuiThemeProvider, createMuiTheme, Button, Grid } from '@material-ui/core'
-import green from '@material-ui/core/colors/green'
+import React, { Component } from 'react'
+import { withStyles, MuiThemeProvider, createMuiTheme, Button } from '@material-ui/core'
+import blue from '@material-ui/core/colors/blue'
+import hash from './hash'
 
 const styles = theme => ({
   button: {
     color: theme.palette,
-    float: 'right',
-    marginRight: 90
+    textDecoration: 'none'
   }
 })
 
 const theme = createMuiTheme({
   palette: {
-    primary: green,
+    primary: blue,
   },
   typography: {
     useNextVariants: true,
   }
 })
 
-function GroceryButton ({ classes }) {
-  return (
-    <MuiThemeProvider theme={theme}>
-      <Grid container>
-        <Grid item xs={12}>
-          <Button className={classes.button}
-            variant="contained"
-            color="primary"
-            >Create Grocery List</Button>
-        </Grid>
-      </Grid>
-    </MuiThemeProvider>
-  )
+class GroceryButton extends Component {
+  constructor(props) {
+    super(props)
+    const { path, params } = hash.parse(location.hash)
+    this.state = {
+      view: { path, params }
+    }
+  }
+  render() {
+    const { classes } = this.props
+    const { id } = this.state.view.params
+    return (
+      <MuiThemeProvider theme={theme}>
+        <Button className={classes.button}
+          variant="contained"
+          color="primary"
+          href={`#grocery-list?id=${id}`}>Create Grocery List</Button>
+      </MuiThemeProvider>
+    )
+  }
 }
 
 export default withStyles(styles)(GroceryButton)
