+++ src/recipe-button.js
@@ -1,0 +1,31 @@
+import React from 'react'
+import { withStyles, MuiThemeProvider, createMuiTheme, Button } from '@material-ui/core'
+import blue from '@material-ui/core/colors/blue'
+
+const styles = theme => ({
+  button: {
+    color: theme.palette
+  }
+})
+
+const theme = createMuiTheme({
+  palette: {
+    primary: blue,
+  },
+  typography: {
+    useNextVariants: true,
+  }
+})
+
+function RecipeButton({ classes }) {
+  return (
+    <MuiThemeProvider theme={theme}>
+      <Button className={classes.button}
+        color="primary"
+        variant="contained"
+        href="#instructions">Make Recipe</Button>
+    </MuiThemeProvider>
+  )
+}
+
+export default withStyles(styles)(RecipeButton)
