+++ src/view-grocery-list.js
@@ -1,4 +1,4 @@
-import React, {Component} from 'react'
+import React, {Component, Fragment} from 'react'
 import GroceryList from './grocery-list'
 import hash from './hash'
